[tests] Fix dtype to np.int

+++ testsuite/parallelization.py
@@ -81,8 +81,8 @@ class TestPythoninterfaceMultiprocessing(unittest.TestCase):
                 line_mapper[iFirst] = tmp
 
                 # Store line segments
-                line_idx += line_mapper[iFirst].tolist()
-                line_step += (i * np.ones(len(iFirst))).tolist()
+                line_idx += line_mapper[iFirst].astype(np.int).tolist()
+                line_step += (i * np.ones(len(iFirst), dtype=np.int)).tolist()
                 line_val += stark_shifted_systems[i].getHamiltonian().diagonal()[iSecond].tolist()
 
                 # Assign the line indices to iSecond, delete line indices that could not be assigned
@@ -159,8 +159,8 @@ class TestPythoninterfaceMultiprocessing(unittest.TestCase):
                 line_mapper[iFirst] = tmp
 
                 # Store line segments
-                line_idx += line_mapper[iFirst].tolist()
-                line_step += (i * np.ones(len(iFirst))).tolist()
+                line_idx += line_mapper[iFirst].astype(np.int).tolist()
+                line_step += (i * np.ones(len(iFirst), dtype=np.int)).tolist()
                 line_val += dipole_interacting_systems[i].getHamiltonian().diagonal()[iSecond].tolist()
 
                 # Assign the line indices to iSecond, delete line indices that could not be assigned
